Feature: Added property write node for devices types and groups
A new node has been added to write a new property of either a device,
type or group, or to modify an existing one.
It allows the filtering of the different assets by their id or name, as
well as the filtering of the properties.
The creation or modification of the properties value is done through an
editor that shows if the JSON is well formed or not.