Feature: Added new asset iterator node and improved server events
This commit also contains many changes regarding the reorganization of
the files of the repository with the objective to make it cleaner.
The whole javascript code regarding the custom options for thinger
assets have been refactored to make it cleaner and easier to use, with
an object oriented programming aproach.